refactor(config): move database related config into own file
