+++ internal/config/config.go
@@ -31,56 +31,6 @@ type Config struct {
 	Warnings []string       `json:"warnings,omitempty"`
 }
 
-// DatabaseProtocol represents a database protocol
-type DatabaseProtocol uint8
-
-func (d DatabaseProtocol) String() string {
-	return databaseProtocolToString[d]
-}
-
-func (d DatabaseProtocol) MarshalJSON() ([]byte, error) {
-	return json.Marshal(d.String())
-}
-
-const (
-	_ DatabaseProtocol = iota
-	// DatabaseSQLite ...
-	DatabaseSQLite
-	// DatabasePostgres ...
-	DatabasePostgres
-	// DatabaseMySQL ...
-	DatabaseMySQL
-)
-
-var (
-	databaseProtocolToString = map[DatabaseProtocol]string{
-		DatabaseSQLite:   "file",
-		DatabasePostgres: "postgres",
-		DatabaseMySQL:    "mysql",
-	}
-
-	stringToDatabaseProtocol = map[string]DatabaseProtocol{
-		"file":     DatabaseSQLite,
-		"sqlite":   DatabaseSQLite,
-		"postgres": DatabasePostgres,
-		"mysql":    DatabaseMySQL,
-	}
-)
-
-type DatabaseConfig struct {
-	MigrationsPath  string           `json:"migrationsPath,omitempty"`
-	URL             string           `json:"url,omitempty"`
-	MaxIdleConn     int              `json:"maxIdleConn,omitempty"`
-	MaxOpenConn     int              `json:"maxOpenConn,omitempty"`
-	ConnMaxLifetime time.Duration    `json:"connMaxLifetime,omitempty"`
-	Name            string           `json:"name,omitempty"`
-	User            string           `json:"user,omitempty"`
-	Password        string           `json:"password,omitempty"`
-	Host            string           `json:"host,omitempty"`
-	Port            int              `json:"port,omitempty"`
-	Protocol        DatabaseProtocol `json:"protocol,omitempty"`
-}
-
 type MetaConfig struct {
 	CheckForUpdates  bool   `json:"checkForUpdates"`
 	TelemetryEnabled bool   `json:"telemetryEnabled"`
@@ -177,19 +127,6 @@ func Default() *Config {
 }
 
 const (
-	// DB
-	dbURL             = "db.url"
-	dbMigrationsPath  = "db.migrations.path"
-	dbMaxIdleConn     = "db.max_idle_conn"
-	dbMaxOpenConn     = "db.max_open_conn"
-	dbConnMaxLifetime = "db.conn_max_lifetime"
-	dbName            = "db.name"
-	dbUser            = "db.user"
-	dbPassword        = "db.password"
-	dbHost            = "db.host"
-	dbPort            = "db.port"
-	dbProtocol        = "db.protocol"
-
 	// Meta
 	metaCheckForUpdates  = "meta.check_for_updates"
 	metaTelemetryEnabled = "meta.telemetry_enabled"
