+++ internal/config/database.go
@@ -1,0 +1,75 @@
+package config
+
+import (
+	"encoding/json"
+	"time"
+)
+
+const (
+	// configuration keys
+	dbURL             = "db.url"
+	dbMigrationsPath  = "db.migrations.path"
+	dbMaxIdleConn     = "db.max_idle_conn"
+	dbMaxOpenConn     = "db.max_open_conn"
+	dbConnMaxLifetime = "db.conn_max_lifetime"
+	dbName            = "db.name"
+	dbUser            = "db.user"
+	dbPassword        = "db.password"
+	dbHost            = "db.host"
+	dbPort            = "db.port"
+	dbProtocol        = "db.protocol"
+
+	// database protocol enum
+	_ DatabaseProtocol = iota
+	// DatabaseSQLite ...
+	DatabaseSQLite
+	// DatabasePostgres ...
+	DatabasePostgres
+	// DatabaseMySQL ...
+	DatabaseMySQL
+)
+
+// DatabaseConfig contains fields, which configure the various relational database backends.
+//
+// Flipt currently supports SQLite, Postgres and MySQL backends.
+type DatabaseConfig struct {
+	MigrationsPath  string           `json:"migrationsPath,omitempty"`
+	URL             string           `json:"url,omitempty"`
+	MaxIdleConn     int              `json:"maxIdleConn,omitempty"`
+	MaxOpenConn     int              `json:"maxOpenConn,omitempty"`
+	ConnMaxLifetime time.Duration    `json:"connMaxLifetime,omitempty"`
+	Name            string           `json:"name,omitempty"`
+	User            string           `json:"user,omitempty"`
+	Password        string           `json:"password,omitempty"`
+	Host            string           `json:"host,omitempty"`
+	Port            int              `json:"port,omitempty"`
+	Protocol        DatabaseProtocol `json:"protocol,omitempty"`
+}
+
+// DatabaseProtocol represents a database protocol
+type DatabaseProtocol uint8
+
+func (d DatabaseProtocol) String() string {
+	return databaseProtocolToString[d]
+}
+
+func (d DatabaseProtocol) MarshalJSON() ([]byte, error) {
+	return json.Marshal(d.String())
+}
+
+const ()
+
+var (
+	databaseProtocolToString = map[DatabaseProtocol]string{
+		DatabaseSQLite:   "file",
+		DatabasePostgres: "postgres",
+		DatabaseMySQL:    "mysql",
+	}
+
+	stringToDatabaseProtocol = map[string]DatabaseProtocol{
+		"file":     DatabaseSQLite,
+		"sqlite":   DatabaseSQLite,
+		"postgres": DatabasePostgres,
+		"mysql":    DatabaseMySQL,
+	}
+)
